Supplement default archive name with classifier extension
Previously, if an archive name was available and did not end in the
extension provided by the classifier, that extension would be added to
aid sflock in identifying the archive type. With this change we also add
the classifier extension to the default archive name used when no
archive name is available to give sflock the same kind of hint in that
case as well.

Signed-off-by: Michael Weiser <michael.weiser@gmx.de>