Extract extension header more robustly
Trying to extract a nonexistant extension header would lead to an
exception and corresponding error message. An empty extension header
would lead to just a trailing dot being added to the archive name. Both
are not fatal but not ideal either. Change the logic to first extract
the header, then add the dot in front and only after that go on with the
comparison. This avoids the exception and error message on nonexistant
header as well.